Move `commSetUnrollFactor()` to rccl_wrap.cc
